More implementation
At the moment everything is kind of everywhere, I'm not happy that I
have to, for example, do relative import (PEP 328) of the custom
execption classes.

Interestingly though, the initial decode/encode classes (extended) seem
to be complete in regard to their very specific purpose. Also, the
settings/config system (thanks to ConfigParser) trivial, but at the
moment I only have a few values...we'll see.

Currently the pymuco script (which will run everything) is a bit of a
mess, I seem to be mixing to many different purposes of code together
thats just making a mess. I'm going to have to sleep on this.